refactor(remoter): 重构remoter，抽离独立模块，职责单一
